pyproject.toml
[tool.poetry]
name = "nonebot-plugin-heweather"
version = "0.6.4"
description = "Get Heweather information and convert to pictures"
authors = ["kexue <x@kexue.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0.0"
httpx = ">=0.18.0,<1.0.0"
nonebot-plugin-htmlrender = ">=0.0.4.3"
nonebot-adapter-onebot = ">=2.0.0-beta.1"
pydantic = ">=1.5.0"
nonebot2 = { version = ">=2.0.0rc1", extras = ["fastapi"] }

[tool.poetry.group.dev.dependencies]
black = { version = "^21.12b0", allow-prereleases = true }
Pillow = "^9.1.1"
isort = "^5.10.1"
nonebug = "^0.2.1"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
